+++ cogment/__init__.py
@@ -26,7 +26,7 @@ framework python SDK by AI Redefined Inc.
 from cogment.version import __version__
 
 from cogment.context import Context, Endpoint, ServedEndpoint
-from cogment.session import EventType
+from cogment.session import EventType, ActorStatus
 from cogment.control import TrialState
 from cogment.datalog_service import LogParams, LogSample
 from cogment.parameters import ActorParameters, TrialParameters