+++ cogment/agent_service.py
@@ -116,19 +116,20 @@ def _process_normal_data(data, session):
         if session._trial.ending and session._auto_ack:
             session._post_outgoing_data(_EndingAck())
 
-        session._trial.tick_id = data.observation.tick_id
+        tick_id = data.observation.tick_id
+        session._trial.tick_id = tick_id
 
         obs_space = session._actor_class.observation_space()
         obs_space.ParseFromString(data.observation.content)
 
         recv_event.observation = RecvObservation(data.observation, obs_space)
-        session._post_incoming_event(recv_event)
+        session._post_incoming_event((tick_id, recv_event))
 
     elif data.HasField("reward"):
         logger.trace(f"Trial [{session._trial.id}] - Actor [{session.name}] received reward")
 
         recv_event.rewards = [RecvReward(data.reward)]
-        session._post_incoming_event(recv_event)
+        session._post_incoming_event((-1, recv_event))
 
         value = recv_event.rewards[0].value
         session._prometheus_data.rewards_counter.labels(session.name, session.impl_name).inc()
@@ -141,7 +142,7 @@ def _process_normal_data(data, session):
         logger.trace(f"Trial [{session._trial.id}] - Actor [{session.name}] received message")
 
         recv_event.messages = [RecvMessage(data.message)]
-        session._post_incoming_event(recv_event)
+        session._post_incoming_event((-1, recv_event))
 
         session._prometheus_data.messages_received.labels(session.name, session.impl_name).inc()
 
@@ -191,7 +192,7 @@ async def _process_incoming(context, session):
                                     f"Trial ended with explanation [{data.details}]")
                     else:
                         logger.debug(f"Trial [{session._trial.id}] - Actor [{session.name}]: Trial ended")
-                    session._post_incoming_event(RecvEvent(EventType.FINAL))
+                    session._post_incoming_event((-1, RecvEvent(EventType.FINAL)))
                 else:
                     if data.HasField("details"):
                         details = data.details